refactor: replace `simp_alive_split` implementation with just simp.
This PR changes the implementation of `simp_alive_split`---which used to call `split` and `split_ifs` and thus introduced new subgoals---with a single simp call (and some `intros` to preserve that behaviour). This is possible as the `if` statements used in the semantics follow a strict pattern, where one branch of the if is always `poison`. Suchs ifs with poison admit straightforward simplification lemmas, which this PR adds, thus removing the need to split on the condition.