To int to nat removal semantics: RISCV Merge [1/3] (#1211)
This is a rather big PR again but it contains the semantics already
merged and additionally has rewrites of the semantics into purely bit
vector domain statements without calls to toInt and toNat. This file
should replace the current smenatics file when complete.

This PR is the first third of the whole undertaking.
For some theorems, we previously already committed their pure variant
but left out the toNat version. In this case, we now add the toNat
version under the original theorem name, move the pure version to a _bv
statement, and prove equivalence of both.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>
Co-authored-by: Tobias Grosser <tobias@grosser.es>
Co-authored-by: Luisa Cicolini <48860705+luisacicolini@users.noreply.github.com>