fix:RISC-V formating (#1232)
This PR adapts to the Lean convention of indenting a match by one
whitespace. Additionally, this PR suppresses unused variable warnings by
prefixing unused variables with an underscore (_). The decision is to
keep these unused variables, particularly those related to RISC-V
operands for clarity. For example, it makes explicit that an addiw
instruction includes an immediate operand as part of its opcode.

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>