fix: adapt RV64 to use bv semantics (#1230)
This PR updates the RiscV64 dialect to use the bit-vector form of each
instruction's semantics. Previously, it relied on semantics involving
many toInt and toNat calls. With this change, the RV64 dialect now
invokes the bit-vector version of each instruction's semantics function.
We also do some minor syntax fixes. In detail, we remove redundant
whitespaces and assign names to unused variables that we still want to
keep for clarity.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>
Co-authored-by: Tobias Grosser <tobias@grosser.es>
Co-authored-by: Luisa Cicolini <48860705+luisacicolini@users.noreply.github.com>