feat: improving-repr-instance-for-pretty-printing-via-opt (#1215)
This PR implements toString functions for the Com and Expr types,
enabling the opt tool to print the bb0.mlir file to the command line in
a more readable format.
It also enhances the opt tool with a --verbose flag, which uses the Repr
instances to print detailed output. If the flag is not provided, the
output defaults to the toString representations.
Currently names are not preserved and the return statement is printed as
a "return" not yet in the dialect generic syntax.
E.g the following input: 
```
  ^bb0(%arg0: i32):
    %0 = "llvm.mlir.constant"() {value = 8 : i32} : () -> i32
    %1 = "llvm.mlir.constant"() {value = 31 : i32} : () -> i32
    %2 = "llvm.or"(%0, %1) <{isDisjoint}> : (i32, i32) -> i32
    %3 = "llvm.sdiv"(%0, %2) <{isExact}> : (i32, i32) -> i32
    "llvm.return"(%3) : (i32) -> ()
```
will be printed to the command line as follows by invoking lake exe opt:
```
^entry(%0 : i32):
   %1 = "llvm.mlir.constant" { value = 8 : i32 } : () → (i32)
   %2 = "llvm.mlir.constant" { value = 31 : i32 } : () → (i32)
   %3 = "llvm.or disjoint" (%1, %2) : (i32, i32) → (i32)
   %4 = "llvm.sdiv exact" (%1, %3) : (i32, i32) → (i32)
  return %4 : (i32) → ()
```

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>
Co-authored-by: Tobias Grosser <tobias@grosser.es>
Co-authored-by: github-merge-queue <118344674+github-merge-queue@users.noreply.github.com>
Co-authored-by: Alex Keizer <alex@keizer.dev>
Co-authored-by: Luisa Cicolini <48860705+luisacicolini@users.noreply.github.com>