fix: change llvm parser to throw unsupported op instead of generic error within mkReturn (#1236)
This PR updates the LLVM parser to throw an "unsupported operation"
error instead of a generic error within the mkReturn function.
Previously, when encountering an unknown return statement or operand, it
would throw a generic error. With this PR, it now throws a more specific
"unsupported op" error.
This PR is part of a stagged PR chain, it is based on the PR #1235.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>