feat: enable opt tool to parse files into LLVMAndRiscV dialect [1/2] (#1242)
This PR adds a new file to the LLVMAndRiscV project folder, which
defines the infrastructure for parsing a region from an input file as a
Com in the LLVMAndRiscV dialect. In future PRs, we will add functions
and flags that enable interaction with the opt tool in the hybrid
dialect setting therefore the suffix [1/2].
This PR also adds an import to the Opt.lean file to allow the opt tool
to access the implemented extensions for the LLVMAndRiscV dialect to
then add the corresponding flags within the opt tool next.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>
Co-authored-by: Alex Keizer <alex@keizer.dev>
Co-authored-by: Tobias Grosser <tobias@grosser.es>
Co-authored-by: github-merge-queue <118344674+github-merge-queue@users.noreply.github.com>