feat: change printing to mirror MLIR syntax closer incl. add formatTypeTupleToString (#1240)
This PR updates the formatting of Com and Expr types to better align
with MLIR syntax conventions. Specifically, it removes unnecessary
whitespace and adjusts the layout to be more consistent with proper MLIR
syntax. For this it introduces a helper function
formatTypeTupleToString, which formats sequences of types by leveraging
existing toString functions.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>