feat: broken simpprocs for the LLVMAndRiscV dialect (#1245)
This PR adds simprocs for the LLVMAndRiscV dialect so that simp_peephole
can recognize and simplify specific patterns. However, the current
simprocs are not yet sufficient. In particular, simplification fails in
the case of cast instructions—we are unable to eliminate the framework
cast, and instead encounter a "move not type correct" error. This issue
may be related to the newly introduced LLVM parser.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>