feat: improve multiPeepholeRewriter (#1277)
This PR updates the multiPeepholeRewriter to support a dependently typed
list of rewrites. This enables handling rewrites with varying contexts
and return types in a single pass. Previously, all rewrites in a list
had to share the same context and return type, requiring multiple calls
to the rewriter for different cases.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>