chore: Verify Simple K-induction algorithm (#1289)
Supercedes https://github.com/opencompl/lean-mlir/pull/1256, and allows
us to prove k-induction step by step. We first prove the "slowest"
version of k-induction in this PR, and prove the faster version in the
subsequent PR.


We depend on 2-3 axioms about the behavior of AIGs, which will be
removed in subsequent PRs.