feat: adding lowerings for constants  (#1282)
This PR adds the lowerings for constants. We use a python script to
generate the rewrites between -25 and 25. Since Lean accepts the proofs,
we have guaranteed correctness for our rewrites no matter how they are
generated.

The script generates the proof of lowering a llvm constant (rhs) to a
RISC-V load immediate instruction (lhs). Below is an example for the
constant one being lowered to a RISC-V load of an immediate with value
one.
```
def llvm_const_lower_riscv_li1 : LLVMPeepholeRewriteRefine 64 [] :=
  {lhs:= const_llvm_1 rhs:= li_riscv_1,
   correct := by
    unfold const_llvm_1 li_riscv_1
    simp_peephole
    simp_riscv
    simp_alive_ops
    simp
  }
```

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>