feat: assemble instruction selection pipeline (#1291)
In this PR, we add all the infrastructure needed to integrate our
instruction selector.
1.) In all modules that define rewriting patterns, we assembled the
rewrites into a dependently typed list. This list will then be accessed
from the instruction selector file and merged into a single list
containing all the rewrites across all instructions.
2.) To simplify this, we added two helper functions that, given a
rewrite, return a dependent pair containing the rewrite along with its
input and output types. This is the exact shape expected by the
rewriter.
3.) This PR also fixes typos found in the trunc and select files.
4.) To complete the setup for instruction selection, this PR adds the
call to the instruction selector from within the Opt tool. For this, we
define a new flag for the Opt tool and specify that the corresponding
action is a call into ParseAndTransform.lean, where the instruction
selection function is invoked.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>