add: basic block extracting script& parsing tests& processing script (#1295)
This PR adds example files to the LLVMAndRISCV test folder that contain
the LLVM IR dialect lowering for each supported instruction in our
fragment. These files were then used to verify that they parts (basic
blocks) can be successfully parsed in Lean-MLIR.
Additionally, this PR adds the following scripts:
A script to extract a basic block from a file.
A script to preprocess an LLVM IR file so that it can be passed to the
opt tool.
A script that performs the full pipeline: preprocessing, passing the
file to opt, and writing the output to a new file.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>