feat: riscv printing& parsing in opt same as in llvm. (#1301)
This PR builds the infrastructure to parse RISC-V using the opt tool. It
also fixes a semantic bug in the RISC-V dialect: the load immediate
instruction (li). This was the only instruction not derived from Sail,
as it is a pseudoinstruction. This PR is very similar to a previous PR
doing the same for llvm and the hybrid dialect.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>