feat: add caching for verified k-induction circuits (#1314)
This attempt uses a different strategy from the previous failed strategy
for caching. We show how to incrementally build our core circuits, and
maintain proofs that our circuits used for recursion are indeed the
outputs of calls to the core circuits, by using a `_succ` theorem that
tells us how to convert a circuit for steps upto `n` to a circuit for
upto steps `n+1`.