feat: Build the k-induction circuit, but by building variables in Circuit for subterm sharing (#1334)
This is much easier to do than changing the whole data structure to AIG,
and should provide similar sharing of key circuits as the AIG encoding
does. While this does not provide *full* subcircuit sharing, it shares
"enough": It shares the computation of the state vector and the output,
which is the key construction that will prevent blowup of the circuit
size.

This is an alternative implementation strategy to #1330 , which requires
moving wholesale to AIGs. That's quite a drastic and deep change, and
would need more time to be robust.