feat: add minimal dockerfile (#1352)
This PR adds a minimal dockerfile for the artifact evaluation. #898 had
a rather large diff with main and included a few changes non strictly
docker-related. However, for the time being, I'd like to start adding
this minimal docker, which I will then update step-by-step as we
finalize the structure of our evaluation.