fix: tweak `hide_constants` to avoid exponentation threshold
This PR tweaks the hide_constants so that `LLVM.const? w x` is now
rewritten to `LLVM.const? (hide w) (hide x)`, instead of `hide
(LLVM.const? _ _)`. This was not necessary to avoid the original deep
recursion errors, but, in some examples it seems that `2 ^ w` was
still being evaluated, leading to errors talking about the
exponentiation threshold. (NOTE: these did not seem to be kernel errors,
rather this seemed to be happening in the elaborator). By explicitly
hiding the `w` value, the problematic expression becomes `2 ^ (hide w)`,
which Lean will not attempt to evaluate, silencing the aforementioned
exponentation threshold error.

We did have to insert a cast, as the return type now changed from `IntW
w` to `IntW (hide w)`, so we had to tweak unhide_constants as well to
properly eliminate this cast.

Furthermore, this PR removes the check whether `x` is a "constant",
rather choosing to hide any expression. There is no real cost to doing
the rewrite, so this isConstant check was unneccessary complexity, which
would've been even more complex to maintain after this change.