feat: refactor scripts for data collection (#1370)
In this PR, I am refactoring the data collection scripts for
`instcombine` and `hackersdelight` such that they rely on common
functions and we can eventually have one single script. I plan to
refactor the `raw-data` folder to have one sub-folder for each
benchmark: to this end, I added the necessary `.placeholder` file, such
that the tracking of their folders is correct.