feat: eliminate counter-examples (#1379)
This PR adds a simproc that rewrites `n % 2 ^ w` into `n` (given that `n
< 2 ^ w`). This simproc has a custom procedure to construct a proof of
the inequality, hence it works smoothly even when the exponent is large
(I tested it for `w = 1000`), without having to touch the exponentiation
threshold. When I ran the evaluation locally, it reported 0
counterexamples :tada:.

This simproc could eventually be upstreamed to bv_decide proper, but
let's first merge it here so that we have a clean artifact.