chore: update and move `collect-stats` (#1386)
This PR updates the `collect-stats` script and its use in CI. 

Changelist: 
* ran `ruff check` and `ruff format` on the `stats` script, removed
unused variables and reformatted, renamed script
* updated CI with the new script name 
* added `tables` subfolder to store the output of the script
* Given that we won't eventually need the `for-paper` subdir (everything
will live in `raw-data`, `plots`, `tables` subdirs for the artifact), I
moved the script out and updated it to save the tables in `tables`. I
also `ruff` formatted and checked the script.
* updated `requirements.txt` with the package the script needs

---------

Co-authored-by: Alex Keizer <alex@keizer.dev>