feat: add handshake-comb dialect (#1448)
This PR adds a new mixed handshake/comb dialect. It also updates
handshake to support streams of bitvecs (as opposed to ints and bools),
while maintaining the infra to potentially add more types if need be. In
particular, I specialized `mkTy` to only work with bitvecs - I could not
find a way to maintain `MkTy2` and parse as a third attribute the
bitvecs' width (which would be the more extensible solution).
Suggestions in this directions are very welcome.