refactor: move term-model to stand-alone file (#1447)
This PR moves the term model development from the core Framework.lean
file to its own file. It seems not to be used anywhere, so I'm not sure
if we want to keep it around, but this is a first step towards
organizing the behemoth that is our Framework.lean file.