WIP: refactor: internalize the denotation of Expr
This PR changes the denotation of Expr to be a map from a Valuation in
context `Γ` to an expanded valuation in context `Γ.snoc t`.
To make this work, we change the "result" index of an Expr from a single
type to the new context, appropriately Forded.

This deduplicates the action of Expr on the
valuation, which is currently defined both in Com and Lets.
Thus, it prepares for supporting multiple return values, where this
action becomes more complex, and thus the deduplication more valuable.