refactor: extract peephole rewriter (#1470)
Stacked on top of #1452 

This PR moves the peephole rewriter out of the monolithic Framework.lean
file and into a separate file for the matcher (matchVar/matchArg) and
the rewriter (rewriteAt and co).

This does break downstream project files, for those I suggest we just
`import SSA.Core`, which will import all of the core development.
This change will make the downstream files resistant to future such
refactors.