refactor: polish EffectKind API (#1492)
This PR cleans up the EffectKind API, in particular by:
1) removing EffectKind.return
2) removing `return_(im)pure_toMonad_eq` simp-lemmas
3) removing `toMonad_(im)pure` from the default simp-set (but keeping it
in simp-denote)

The first is easy: `EffectKind.return` was never used, we always write
(`Pure.`)`pure`.

The reason for removing `return_pure_toMonad_eq` in particular, is that
unfolding `Id x` to `x`, and especially rewriting `(pure y : Id _)` to
just `y` is now considered bad upstream, and simp-lemmas that did so
have been deprecated. This is essentially the same as
`return_pure_toMonad_eq` would do, so we remove it to follow suit.

As a consequence, we've had to fix proofs that used. Those that actually
relied on the def-eq still go through with an explicit `rfl`, but as it
turns out, most proofs didn't really need to unfold this, and I've
simplified most proofs that used it.