refactor: extract `Mapping` type to separate file, clean up proofs (#1502)
This PR moves the `Mapping` from the megalithic Framework file into a
separate file.
Furthermore, we add a `Mapping.toHom` function which converts a partial
Mapping together with a proof that it is actually total into a
`Ctxt.Hom`

We then make a pass over the matchVar &co proofs, simplify them to use
Ctxt.Hom and generally clean them up.