refactor: make Com.rec' recursor more flexible (#1510)
This PR changes the Com.rec' custom recursor for simple structural
recursion on `Com`s (i.e., without mutually recursion on the contained
Exprs) more flexible.

Specifically, it ensures the effect and return type--which never
actually change--are treated as parameters rather than index. This means
we'll be allowed to do `induction com` even when the effect or return
type are specific expressions (rather than free variables).

This also means there's no need for `recPure` anymore, which previously
we had to use to be able to recurse on known-pure `Com`s; this now
becomes just an instance of the more flexible recursion principle.