chore: peel off TacBench
This separates out the two components needed to run QF_BV goals
separately: the parametric bitvector decision procedure, and the
tac_bench tactic. Together, this permits us to run our evaluation
without the full `lean-mlir` being on the critical path of the build,
which will help with CI runs feeling fast at:

https://github.com/opencompl/evaluation-multi-width-bv/actions/runs/17154834586/job/48669126582

As usual, things "feeling fast" is important for developer productivity
and happiness, which I am attempting to maximize.