feat: add `ToPrint` to framework for better MLIR printing (#1539)
The current `toString` implementation is not sufficiently generic to
print correct MLIR syntax. This PR introduces a new `toPrint`
infrastructure that enables accurate printing of MLIR syntax.

We should investigate if and how much of the `toString` infrastructure
is still needed and eliminate this redundancy.

This PR imports the changes to `Framework.lean`, to `Base.lean`, and to
`Opt.lean` from #1538.

Authored by @salinhkuhn.

---------

Co-authored-by: Tobias Grosser <tobias@grosser.es>