chore: refine options for large Op inductive in RISCV64 (#1556)
Given that there is only one inductive that needs more heartbeats and
recursion depth, this change moves to a local `set_option ... in`.

In preparation of upcoming changes, we also increase the stack-space
usage of the lean server to avoid crashes of VScode as we will add more
constructors to the inductive.