feat: `LLVMAndRiscV` refactoring (#1543)
This PR refactors the existing hybrid dialect (that includes LLVM and
RISCV), implementing our convention and checking comments and style. I
renamed the hybrid dialect to `LLVMPlusRiscV` (it previously was
sometimes referred to as `LLVMAndRiscV` or `LLVMRiscV`: this is now
consistent). I would like to rename everything to `LLVMAndRiscV`, to be
consistent with the Lean-MLIR costume of calling the file and the
dialect the same, but this is not a priority at the moment and I will do
that in a subsequent PR.

authored by @salinhkuhn