fix: properly canonicalize types in `reduceIsRefinedBy` (#1594)
In particular, when unfolding `inferInstance(As)`, rewrite the goal to
be in terms of the *inferred* types, instead of the types originally
present in the goal.

That is, suppose refinement for `⟦.bitvec w⟧` is defined via
`inferInstanceAs` and `LLVM.IntW w`, then this PR will ensure a goal
that has `@IsRefinedBy ⟦.bitvec w⟧ ...` will be rewritten to
`@IsRefinedBy (LLVM.IntW w)`.