feat: Add support for multiple return values (#1511)
This PR adds support for multiple return values in `Com`!

The core change is that `DialectSignature.outTy` now becomes
`DialectSignature.returnTypes`, and is generalized to return a List of
types. All the other changes are effectively the natural consequence of
this.

- The proof strategy in `Match.lean` has changed to rely more heavily on
Subtypes. This means that we can have simp lemmas that are effectively
keyed on the goals, instead of having to rely on remembering which
rewrite to apply (since they are disambiguated by side-condition).
- DCE had to be generalized to delete a (consecutive) range of variables
- We drive-by fix some printing instances so that (a) HVector's printing
no longer includes stray extra brackets and commas and (b) so that the
Com/Expr printing code does not include brackets/list-notation, but
instead look like, e.g., `(i64, i64)`
- The TermModel code got broken, but it's not actually used anywhere, so
:shrug:
- Finally, we adjust the entire rest of the project to align with the
possibility of multiple return values:
- Dialect operation semantics now have to return an hvector of values
(since there could be multiple); given that existing dialects only have
a single return for each operation, we just wrap the semantics in
`[...]\_h`
- Similarly, we have to introduce some list literals whenever a `Com` or
`Expr` type is referenced:
- There is also a `LeanMLIR.SingleReturnCompat` namespace which contains
backwards compatible aliasses, so for simple cases a straightforward fix
is `open LeanMLIR.SingleReturnCompat` to recover the old behaviour.
- Finally, we had to fix the proof automation to deal with the new shape
of semantics:
- In particular, I wrote `simpValuationApply`, which is a (d)simproc
that normalizes "applications of *fully-concrete* valuations to
*fully-concrete* variables", so for example, `(\empty ::\_v x ::\_v y)
\<1, _\>` is rewritten to just `x`.
- This used to be done by composing simp lemmas and hoping for the best;
this somehow broke but rather than figuring out why exactly moving to a
simproc gives us much more control and thus robustness