feat: run evaluation steps concurrently (#1599)
Split off the building of Lean into a separate step that caches its
outputs, then extract three evaluation steps: `Check extracted goals`,
`Evaluate LLVM`, `Evaluate Alive & Hacker's Delight`. These evaluation
steps can run concurrently, all reusing the previous Lean-MLIR build. We
build Mathlib from scratch (not using the pre-built Mathlib) as this
allows us to build only the files in Mathlib that we actually use, which
reduces the size of the cache we store on GitHub.

We use GitHub's `action/cache` to pass the build data from the `Build
Lean-MLIR` step to the individual evaluation steps, which is fast and
seems less complex than building an artifact. When the cache is hit, the
`Build Lean-MLIR` step takes 7 seconds. Our full cache with Lean-MLIR,
Mathlib, and Lean requires 1200 MB and can be downloaded in 16 seconds.

This reduces the runtime of our evaluation from 45 minutes to 20
minutes.