just bake the .git history in the Docker image
Yes, this will cause more (layer) cache misses, but our CI actions actually use the git history so it is needed, and trying to recover the history using CI actions is not great. Most diffs will change the code anyway, and we will still have the cache mount of .lake, so it should be fine.

If not, we can always optimize the layering