feat: CI-checked MathlibImports.lean
This PR adds a new file, MathlibImports.lean, which imports all Mathlib
files that are used anywhere in the project, and adds a CI action that
ensures this file is kept up-to-date.

We then use this file in the Docker build to cache only those Mathlib
files we actually use, minimizing the size of the cache and thus
speeding up subsequent builds (at the cost of building Mathlib
ourselves, rather than using `lake exe cache get`, since the latter
will indiscriminately download all of Mathlib).