feat: duplicate evaluation CI to run in docker as well (#1625)
As per #1613, I'm experimenting with an alternative CI setup where we
run everything inside a Docker container. Until we commit to this being
the workflow we like, I'll duplicate existing CI jobs and have a
*seperate* job run the same action inside the container (on a
GH-provided runner, but using the image we built on namespace's runner).

This PR does so for the CI job that builds the non-standard build
targets and the Instcombine evaluation (but not yet for hackersdelight,
for that I'm waiting on #1549).

To make this work, I:
- Add more stuff to the .dockerignore file, so files which are not
needed are not copied to the Docker image (and thus won't trigger cache
misses).
- Expand the existing bitwuzla image to also install uv, use uv to
install the python dependencies specified in requirements.txt, and copy
the lean-mlir files from the base-image to get a full-blown instcombine
image; the produced image is thus renamed to `lean-mlir-instcombine`.
Note that this second image is still build on GH runner infrastructure!
This is fine as we don't actually `lake build` anything here, so there
is no need for incremental caching (but of course, using namespace.so
would still likely be faster).
- I tried dropping permissions in the Dockerfile to use a non-root user
- Github will change the home-directory to `/github/home`, so I
originally thought it was using a github user because of security
concerns with using a root user. Thus, I tried to set a `USER` in the
dockerfile.
- However, it just made everything more complicated, and random actions
started failing with permission errors.
- Then I tried setting the home-directory explictly to `/root` using
`env.var`, but then actions started complaining about
`/root/.docker/config.json` not existing...
- Just leaving the status quo is non-ideal, as .elan was re-downloading
the toolchain (since the Docker image has it at /root/.elan, while in
the action it was looking at `/github/home/.elan`
- Finally, I worked around it by adding a first run step that symlinks
the latter to the former to every job that runs in a container. It's not
the most elegant, as it adds a boilerplate run step to every job, but it
actually works, so :shrug: