refactor: rename `{Ctxt,Valuation}.snoc` to `cons`
This PR changes contexts and valuations to be proper cons-lists.
Previously, we'd named their "add-one-element" constructor `snoc`, but
then we implemented this in terms of `List.cons` and generally were
quite inconsistent in e.g., the append and getElem methods.

In fact, besides the name all other aspects seemed to behave as a
snoc-list, so I decided to pick the snoc-list perspective as canonical.