feat: add (very) basic support for multiple returns in `def_signature`.
The new code checks whether the rhs of the arrow in the def_denote syntax is of type `«term[_]»`; this ought to work for a straight list literal, but is likely super fragile in the presence of anything else, even just `(...)` parentheses around the literal. There is not really a good way to avoid that given the current syntax-rewriting based setup.