feat: [Blase] add support for predicate variables (#1657)
This ensures that when we see a goal state we cannot fully handle (e.g.
presence of `mulOverflow`) the decision procedure does not give up
entirely. Instead, it abstracts the proposition out into a new
"proposition variable" (exactly as we do for BV expressions that are
outside of our domain) and tries to solve the problem anyway. This
should provide more accurate error messages, and allows us to understand
which Prop fragments the tactic is currently unable to handle, making
improving our numbers (and preprocessing) much easier.