chore: add a parse and print test-case for each LLVM operation (#1670)
I've generated a test-case for each operation in the LLVM dialect, where
we parse a program in the EDSL that uses (only) that operation and print
it using the various printing methods.

This currently exposes many bugs in the printing, but I'm first
committing all test-cases so that we can use it to track progress on
fixing those bugs.

NOTE: the test-cases were generated using an LLM; it's actually one of
the few times it turned out to be good at generating Lean code (although
it certainly still did some dumb stuff, too, in the process)